Move ConnectionForm out into own file
